commenting finished

+++ ETick/src/main/java/uk/ac/cam/cl/ticking/ui/api/UserApiFacade.java
@@ -24,6 +24,7 @@ import uk.ac.cam.cl.ticking.ui.configuration.ConfigurationLoader;
 import uk.ac.cam.cl.ticking.ui.dao.IDataManager;
 import uk.ac.cam.cl.ticking.ui.ticks.Fork;
 import uk.ac.cam.cl.ticking.ui.ticks.Tick;
+import uk.ac.cam.cl.ticking.ui.util.Strings;
 
 import com.google.inject.Inject;
 
@@ -61,7 +62,17 @@ public class UserApiFacade implements IUserApiFacade {
 	public Response getUser(HttpServletRequest request) {
 		String crsid = (String) request.getSession().getAttribute(
 				"RavenRemoteUser");
+		
+		/* Get the user object, returning if not found */
 		User user = db.getUser(crsid);
+
+		if (user == null) {
+			log.error("Requested user " + crsid
+					+ " but they couldn't be found");
+			return Response.status(Status.NOT_FOUND).entity(Strings.MISSING)
+					.build();
+		}
+		
 		return Response.ok(user).build();
 	}
 
@@ -72,6 +83,17 @@ public class UserApiFacade implements IUserApiFacade {
 	public Response deleteUser(HttpServletRequest request, String crsid,
 			boolean purge) {
 		// TODO admin check
+		
+		/* Get the user object, returning if not found */
+		User user = db.getUser(crsid);
+
+		if (user == null) {
+			log.error("Requested user " + crsid
+					+ " for deletion, but they couldn't be found");
+			return Response.status(Status.NOT_FOUND).entity(Strings.MISSING)
+					.build();
+		}
+		
 		db.removeUser(crsid, purge);
 		return Response.ok().build();
 	}
@@ -130,11 +152,24 @@ public class UserApiFacade implements IUserApiFacade {
 	public Response addSSHKey(HttpServletRequest request, String key) {
 		String crsid = (String) request.getSession().getAttribute(
 				"RavenRemoteUser");
+		
+		/* Get the user object, returning if not found */
+		User user = db.getUser(crsid);
+
+		if (user == null) {
+			log.error("Requested user " + crsid
+					+ " to add a public ssh key, but they couldn't be found");
+			return Response.status(Status.NOT_FOUND).entity(Strings.MISSING)
+					.build();
+		}
+		
+		/*Call the git service*/
 		try {
 			gitServiceProxy.addSSHKey(key, crsid);
 		} catch (InternalServerErrorException e) {
 			RemoteFailureHandler h = new RemoteFailureHandler();
 			SerializableException s = h.readException(e);
+			
 			log.error("Tried adding ssh key for " + crsid, s.getCause());
 			return Response.status(Status.INTERNAL_SERVER_ERROR).entity(e)
 					.build();
@@ -143,7 +178,8 @@ public class UserApiFacade implements IUserApiFacade {
 			return Response.status(Status.INTERNAL_SERVER_ERROR).entity(e)
 					.build();
 		}
-		User user = db.getUser(crsid);
+		
+		/*Add the key to the user object, save and return*/
 		user.setSsh(key);
 		db.saveUser(user);
 		return Response.status(Status.CREATED).entity(user).build();
